Public API

EmpiricalModeDecomposition.ceemd — Method

ceemd(measurements, xvec; num_imfs=6)

Compute the Complete Empirical Mode Decomposition of the time series with values measurements and time steps xvec. numimfs is the Number of Intrinsic Mode Functions. Returns a list of numimfs + 1 Vectors of the same size as measurements.

EmpiricalModeDecomposition.eemd — Function

eemd(measurements, xvec, numtrails=100)

Return the Intrinsic Mode Functions and the residual of the ensemble Empirical Mode Decomposition of the measurements given on time steps xvec.

EmpiricalModeDecomposition.emd — Function

emd(measurements, xvec)

Return the Intrinsic Mode Functions and the residual of the Empirical Mode Decomposition of the measurements given on time steps given in xvec.

EmpiricalModeDecomposition.maketestdata — Method

maketestdata(seed)

Return a simple example time series with the composing parts

EmpiricalModeDecomposition.EMDIterable — Type

Iterator for the Empirical Mode Decomposition. The time series values are an AbstractVector of type T, and the time positions are an AbstractVector of type U.

EmpiricalModeDecomposition.SiftIterable — Type

SiftIterable{T, U}

Iterator for the sifting algorithm. The time series values are an AbstractVector of type T, and the time positions are an AbstractVector of type U. Fields:

Internal API

EmpiricalModeDecomposition.colominas2014_s — Method

colominas2014_s()

First testdata from Colominas 2014 et al. http://dx.doi.org/10.1016/j.bspc.2014.06.009

EmpiricalModeDecomposition.colominas2014_x — Method

colominas2014_x()

Second testdata from Colominas 2014 et al. http://dx.doi.org/10.1016/j.bspc.2014.06.009 It is not exactly the same, because ϕ is not feasable because the arccos is only defined for values between -1 and 1.

EmpiricalModeDecomposition.fosso — Function

fosso2014()

Make testdata from Fosso 2014 et al.

EmpiricalModeDecomposition.get_edgepoint — Method

get_edgepoint(y, xvec, extremas, pos, comp)

Compute the edgepoint which should be used as the extrema on the edge for the spline computation.

EmpiricalModeDecomposition.ismonotonic — Method

ismonotonic(x::AbstractVector)

Check wether x is monotonic. This means, every value is either larger or smaller than the preceding value.

EmpiricalModeDecomposition.localmaxmin! — Method

localmaxmin!(x, maxes, mins)

Detect the local extrema of x. Push the maxima into maxes and the minima into mins.

EmpiricalModeDecomposition.sift — Function

sift(y, xvec)

Sift the vector y whose points have x coordinates given by xvec.

EmpiricalModeDecomposition.zerocrossing! — Method

zerocrossing!(y, crosses)

Compute the indices of zerocrossings of a vector It searches for elements which are either zero or near a signflip and pushes the indices into crosses.

EmpiricalModeDecomposition.CEEMDIterable — Type

Iterator for the Complete Empirical Mode Decomposition. The time series values are an AbstractVector of type T, and the time positions are an AbstractVector of type U.

EmpiricalModeDecomposition.CEEMDState — Type

Intermediate results of the CEEMD Iteration

EmpiricalModeDecomposition.SiftState — Type

SiftState

Handle the intermediate results of the sifting.