Extras

Extra Functions

EmpiricalModeDecomposition.hilbert_transform — Method.

hilbert_transform(signal::Vector{Float64})

Compute the Hilbert transform using the DFT approximation.

source

EmpiricalModeDecomposition.hht — Method.

hht(signal::Vector{Float64}, s::EMDSetting)

Return the instantaneous energies and frequencies of the IMFS computed via the Hilbert-Huang transform with EMD settings specified.

source

EmpiricalModeDecomposition.hht — Method.

hht(signal::Vector{Float64}, s::EEMDSetting)

Return the instantaneous energies and frequencies of the IMFS computed via the Hilbert-Huang transform with EEMD settings specified.

source