Internals

Internal functions

EmpiricalModeDecomposition.sift! — Method.

sift!(input::Vector{Float64}, s::EMDSetting)

Return the IMF that satisfies the given settings. In particular, it returns an IMF that satisifes the S number criterion that is found within the set number of siftings.

source

EmpiricalModeDecomposition.find_extrema! — Method.

find_extrema!(x::Vector{Float64}, max_x::Vector{Float64}, max_y::Vector{Float64},
              min_x::Vector{Float64}, min_y::Vector{Float64})

Return the number of maxima, minima, and zero crossings of x after modifying maxx, maxy, minx, miny to contain the maxima and minima of x.

source

EmpiricalModeDecomposition.linear_extrapolate — Method.

linear_extrapolate(x_0::Float64, y_0::Float64, x_1::Float64, y_1::Float64, x::Int64)

Return the linear extrapolation of x based on x0, x1, y0, y1.

source

EmpiricalModeDecomposition.evaluate_spline — Method.

evaluate_spline(x::Vector{Float64}, y::Vector{Float64}, n::Int64)

Return spline generated by the first n elements of (x,y).

source