Library

Public API

EmpiricalModeDecomposition.ceemd — Method.

ceemd(measurements, xvec; num_imfs=6)

Compute the Complete Empirical Mode Decomposition of the time series with values measurements and time steps xvec. numimfs is the Number of Intrinsic Mode Functions. Returns a list of numimfs + 1 Vectors of the same size as measurements.

source

EmpiricalModeDecomposition.eemd — Function.

eemd(measurements, xvec, numtrails=100)

Return the Intrinsic Mode Functions and the residual of the ensemble Empirical Mode Decomposition of the measurements given on time steps xvec.

source

EmpiricalModeDecomposition.emd — Function.

emd(measurements, xvec)

Return the Intrinsic Mode Functions and the residual of the Empirical Mode Decomposition of the measurements given on time steps given in xvec.

source

EmpiricalModeDecomposition.maketestdata — Method.

maketestdata(seed)

Return a simple example time series with the composing parts

source

EmpiricalModeDecomposition.EMDIterable — Type.

Iterator for the Empirical Mode Decomposition. The time series values are an AbstractVector of type T, and the time positions are an AbstractVector of type U.

source

EmpiricalModeDecomposition.SiftIterable — Type.

SiftIterable{T, U}

Iterator for the sifting algorithm. The time series values are an AbstractVector of type T, and the time positions are an AbstractVector of type U. Fields:

source

Internal API

EmpiricalModeDecomposition.get_edgepoint — Method.

get_edgepoint(y, xvec, extremas, pos, comp)

Compute the edgepoint which should be used as the extrema on the edge for the spline computation.

source

EmpiricalModeDecomposition.ismonotonic — Method.

ismonotonic(x::AbstractVector)

Check wether x is monotonic. This means, every value is either larger or smaller than the preceding value.

source

EmpiricalModeDecomposition.localmaxmin! — Method.

localmaxmin!(x, maxes, mins)

Detect the local extrema of x. Push the maxima into maxes and the minima into mins.

source

EmpiricalModeDecomposition.sift — Function.

sift(y, xvec)

Sift the vector y whose points have x coordinates given by xvec.

source

EmpiricalModeDecomposition.zerocrossing! — Method.

zerocrossing!(y, crosses)

Compute the indices of zerocrossings of a vector It searches for elements which are either zero or near a signflip and pushes the indices into crosses.

source

EmpiricalModeDecomposition.CEEMDIterable — Type.

Iterator for the Complete Empirical Mode Decomposition. The time series values are an AbstractVector of type T, and the time positions are an AbstractVector of type U.

source

EmpiricalModeDecomposition.CEEMDState — Type.

Intermediate results of the CEEMD Iteration

source

EmpiricalModeDecomposition.SiftState — Type.

SiftState

Handle the intermediate results of the sifting.

source